requirements.txt
numpy==1.22.3
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pygame==2.1.2